LCI_ChatBot/requirements_core.txt
# Core PDF parsing dependencies - Python 3.12 Compatible
# Essential packages for PDF parsing functionality

# Core PDF processing
PyMuPDF==1.24.10
pdfplumber==0.11.4

# Data processing
numpy==1.26.4
pandas==2.2.3
requests==2.32.3

# Text processing
tiktoken==0.7.0
regex==2024.9.11

# Optional PDF processing (install if needed)
unstructured[pdf]==0.15.12
camelot-py==0.11.0
tabula-py==2.9.3

# Optional ML/AI (install if needed)
torch==2.5.1
sentence-transformers==3.2.1
transformers==4.46.2

# Optional utilities
python-dotenv==1.0.1
tqdm==4.66.5
matplotlib==3.9.2
seaborn==0.13.2

# OCR (optional)
pytesseract==0.3.13
Pillow==10.4.0
opencv-python==4.10.0.84
pdf2image==1.17.0
